add class Series and Catalogo in correct project
