+++ Catalogo/Catalogo/classSeries.swift
